Make scanner accept dot as standalone symbol
Currently PL/Proxy fails to parse following query:

  select (ret_numtuple(1)).num, (ret_numtuple(1)).name;

The fix is to add "." to the SQLSYM rule in the lexer.  This should be OK
because an identifier chain will always be longer than a single dot, so flex
will prefer it over parsing the single dot.

Patch by Peter Eisentraut