Add language validator
Refactor some internals to make this possible.  Mainly, FunctionCallInfo
is not available when validating, so avoid accessing that if not
required.  Rename plproxy_compile() to plproxy_compile_and_cache() and
the previously internal fn_compile() to plproxy_compile().  This matches
their purpose better and allows the validator to call plproxy_compile()
without invoking execution-time dependent code.

Many error test cases have changed because the validator catches errors
when the function is created, not when it is called.

Raise the extension version to 2.5.1 to be able to upgrade from
non-validator installations.