Fix yy_scan_bytes() prototype
Flex changed the second argument of yy_scan_bytes() from int to size_t a
great while ago, while not updating the documentation.  To make the
prototype in plproxy match, use flex --header-file to create a header
file, and include that, instead of maintaining an explicit prototype
ourselves.

See also:

- <https://sourceforge.net/p/flex/bugs/184/>
- <https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=750163>