Enable column name mapping for set returning functions
This requires moving result_map from ProxyFunction to ProxyConnection,
so that we can have one per connection.

That was not necessary before, because we got all results from one connection
in a single PGresult, and we could afford to initialize the mapping whenever
we got a result.

Now, with single-row mode enabled, the result rows don't arrive connection
by connection, so we have to maintain one result_map per connection, because
setting up the mapping for every single result would be way too expensive.

The result_map is now allocated in the persistent cluster_mem context, so
we have to take care to free it whenever the ProxyConnection is cleaned
or freed.

In passing, rearrange the code in plproxy_tuple_from_result to have less
indentation.

Add a regression test for the column name mapping functionality that was
previously missing.