Handle result sets with a tuplestore
Before, the whole result set was cached on memory, which can
easily lead to an out-of-memory condition for big result sets.

Instead, retrieve query results in single row mode and use a
tuplestore for intermediate storage until the result is complete.
Tuplestores automatically spill to temporary files if they get
too large.

This can change the order of results, so add an ORDER BY to
the regression tests for stability.

This breaks column name mapping, because results are not processed
connection by connection now.
This will be fixed in the next commit.